requirements.txt
Django
djangorestframework
django-filter
psycopg2-binary
python-dotenv
drf-yasg